Bean Scopes singleton vs prototype
Bean Scopes:

there are mainly two scopes on beans i.e singleton and prototype. there are scopes like request, session global session. But they are only useful for web apps.

By default bean scope is singleton. Means (a little recap of singleton) there would only one instance per bean. Example. In DemoController there is a request named check. It check if the two objects  defined in demoController refer to the same bean or not. So when  you ll run the application, and hit the check request, it will show u 'true' because default scope is singleton.

Prototype:
In prototype scope, a new bean is initialized for every request. Example: To see this in action, uncomment the Scope the line in CricketCoach, and re run the application and hit check url. you ll see now it will return the false value which means now mycoach and anotherCoach are separate beans