Core Demo: Beans Vs Components
Both beans and Component annotation used for creating a spring bean class. Then question arises why we need 2 of them. though both of them provides us with the spring beans, but their usage differentiate it.

So far all the code that has been done in this project, we were using component annotation  for creating the bean. But now lets imagine for a second that there is a class that you want use as a spring bean but its in 3rd party library. So now you can't add the use the component annotation because its in the 3rd party lib. To overall this situation, comes the bean annotation. It helps us to use those classes as bean which are not originally part of your project or don't have access to edit them and add component annotation. So to add this class in your spring create a class with configuration annotation. example SportConfig. and create a method to instantiate the object of the class you want to include in your spring project. and thats it. this class is now part of your spring project.

 this is the main difference. Other than that, bean annotation use at method while component annotation use at class.